passph.py: Add --alt-charlist option to use alternative charlist
This option was added to either obtain a charlist from the environment
variable PASSPH_CHARLIST or a preset builtin. The builtin only provides
62 unique characters, which are only [a-zA-Z0-9], as this still seems to
be the best bet for a list of characters for services which only accept
limited range of characters. Not particularly happy about having to
revert to a reduced character set of 62 chars, but in 2013 this still
appears to be a necessity more often than I'd like.

Signed-off-by: Marco Elver <me@marcoelver.com>